2da entrega CRUD completo

+++ src/main/java/com/example/demo/repository/VentaRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.modelo.Venta;
 
-public interface VentaRepository extends JpaRepository<Venta, Long> { }
+public interface VentaRepository extends JpaRepository<Venta, Integer> { }
